+++ src/babel/__tests__/examples/stylesheet_empty.js
@@ -1,0 +1,15 @@
+import * as zacs from '@nozbe/zacs'
+
+const styles = zacs.stylesheet({
+  empty: {},
+  emptyOnWeb: {
+    native: {
+      backgroundColor: 'red',
+    },
+  },
+  emptyOnNative: {
+    web: {
+      width: 100,
+    },
+  },
+})